requirements.txt
asgiref==3.5.2
boto3==1.24.78
botocore==1.27.78
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-oauth-toolkit==2.1.0
django-phonenumber-field==7.0.0
django-storages==1.13.1
djangorestframework==3.13.1
drf-social-oauth2==1.2.1
drf-yasg==1.21.3
ecdsa==0.18.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
jwcrypto==1.4.2
Markdown==3.4.1
MarkupSafe==2.1.1
oauthlib==3.2.1
packaging==21.3
phonenumbers==8.12.55
Pillow==9.2.0
psycopg2==2.9.3
pyasn1==0.4.8
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
twilio==7.13.0
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.2.0
wrapt==1.14.1
